Squashed commit of the following:
commit f65a45cdb4ee72acebd956b11dce00c4f2fd7810
Merge: 51fd452 08d0f2a
Author: ericthered1138 <ericjennings955@gmail.com>
Date:   Fri Jan 21 09:55:38 2022 -0600

    Merge remote-tracking branch 'origin/dZazulak/python/profile_create_post/sp1' into eJennings/python/userprofile/sprint1

commit 08d0f2a3535d8e30f0accb4c5790c2fd03751bde
Author: dZazulak <94470265+dZazulak@users.noreply.github.com>
Date:   Fri Jan 21 09:54:17 2022 -0600

    Implemented simple test/DAO method for update profile

commit bc98d55eded0cd81acc928488029b02d5b6502d4
Author: dZazulak <94470265+dZazulak@users.noreply.github.com>
Date:   Fri Jan 21 09:17:09 2022 -0600

    Squashed commit of the following:

    commit 51fd452a9ffa28086278d08dce1645d71296476f
    Author: ericthered1138 <ericjennings955@gmail.com>
    Date:   Thu Jan 20 23:50:43 2022 -0600

        fixed bad input error

    commit cdd161d629971446acda61870a2c984210ab07f7
    Author: ericthered1138 <ericjennings955@gmail.com>
    Date:   Thu Jan 20 23:34:25 2022 -0600

        finished image methods

    commit d1367bae6115ac8e28c6171ac4b2b51901517975
    Author: ericthered1138 <ericjennings955@gmail.com>
    Date:   Thu Jan 20 22:32:13 2022 -0600

        Update main.py

    commit 92b70d8b5cac73202d78c02170525809a6512787
    Author: ericthered1138 <ericjennings955@gmail.com>
    Date:   Thu Jan 20 22:24:07 2022 -0600

        Update test_create_post_service.py

    commit 92caa00cf4a25d4411a478c29c6373ef49e7d612
    Author: ericthered1138 <ericjennings955@gmail.com>
    Date:   Thu Jan 20 15:51:07 2022 -0600

        Update test_create_post_service.py

    commit 6ecdb922230b0119c6ace41be1558fc80e4eecd7
    Author: ericthered1138 <ericjennings955@gmail.com>
    Date:   Thu Jan 20 14:08:48 2022 -0600

        finished data access service layer in progress

    commit de9a8e81a04f9a903aa8a99cebccae7c97824558
    Author: ericthered1138 <ericjennings955@gmail.com>
    Date:   Thu Jan 20 11:11:30 2022 -0600

        fixed abstract methods for user_profile

    commit bc23fa372b252059fbf47d5fae6b0b0da0189883
    Author: ericthered1138 <ericjennings955@gmail.com>
    Date:   Wed Jan 19 23:49:17 2022 -0600

        added some image data access methods

    commit 5fd64ac15d51f02a0cb4e94f4b2d2d3704e51960
    Author: ericthered1138 <ericjennings955@gmail.com>
    Date:   Wed Jan 19 22:29:57 2022 -0600

        added insert image method

    commit 5bddca3bbcacf9a5873acef4bce843f6ad1ed18f
    Author: ericthered1138 <ericjennings955@gmail.com>
    Date:   Wed Jan 19 16:11:30 2022 -0600

        1

    commit 3b7f8a38eb2ec1b694a229c4e143df8c41a7b8d9
    Author: ericthered1138 <ericjennings955@gmail.com>
    Date:   Wed Jan 19 13:26:28 2022 -0600

        updated layer layout

commit 51fd452a9ffa28086278d08dce1645d71296476f
Author: ericthered1138 <ericjennings955@gmail.com>
Date:   Thu Jan 20 23:50:43 2022 -0600

    fixed bad input error

commit cdd161d629971446acda61870a2c984210ab07f7
Author: ericthered1138 <ericjennings955@gmail.com>
Date:   Thu Jan 20 23:34:25 2022 -0600

    finished image methods

commit d1367bae6115ac8e28c6171ac4b2b51901517975
Author: ericthered1138 <ericjennings955@gmail.com>
Date:   Thu Jan 20 22:32:13 2022 -0600

    Update main.py

commit 92b70d8b5cac73202d78c02170525809a6512787
Author: ericthered1138 <ericjennings955@gmail.com>
Date:   Thu Jan 20 22:24:07 2022 -0600

    Update test_create_post_service.py

commit 92caa00cf4a25d4411a478c29c6373ef49e7d612
Author: ericthered1138 <ericjennings955@gmail.com>
Date:   Thu Jan 20 15:51:07 2022 -0600

    Update test_create_post_service.py

commit 6ecdb922230b0119c6ace41be1558fc80e4eecd7
Author: ericthered1138 <ericjennings955@gmail.com>
Date:   Thu Jan 20 14:08:48 2022 -0600

    finished data access service layer in progress

commit de9a8e81a04f9a903aa8a99cebccae7c97824558
Author: ericthered1138 <ericjennings955@gmail.com>
Date:   Thu Jan 20 11:11:30 2022 -0600

    fixed abstract methods for user_profile

commit bc23fa372b252059fbf47d5fae6b0b0da0189883
Author: ericthered1138 <ericjennings955@gmail.com>
Date:   Wed Jan 19 23:49:17 2022 -0600

    added some image data access methods

commit 5fd64ac15d51f02a0cb4e94f4b2d2d3704e51960
Author: ericthered1138 <ericjennings955@gmail.com>
Date:   Wed Jan 19 22:29:57 2022 -0600

    added insert image method

commit 5bddca3bbcacf9a5873acef4bce843f6ad1ed18f
Author: ericthered1138 <ericjennings955@gmail.com>
Date:   Wed Jan 19 16:11:30 2022 -0600

    1

commit 3b7f8a38eb2ec1b694a229c4e143df8c41a7b8d9
Author: ericthered1138 <ericjennings955@gmail.com>
Date:   Wed Jan 19 13:26:28 2022 -0600

    updated layer layout

+++ PythonAPI/data_access_layer/implementation_classes/user_profile_dao_imp.py
@@ -1,3 +1,5 @@
+from custom_exceptions.user_image_not_found import UserImageNotFound
+from custom_exceptions.user_not_found import UserNotFound
 from data_access_layer.abstract_classes.user_profile_dao import UserProfileDAO
 from entities.user import User
 from util.database_connection import connection
@@ -6,20 +8,27 @@ from util.database_connection import connection
 class UserProfileDAOImp(UserProfileDAO):
 
     def get_user_profile(self, user_id: int) -> User:
-        """Grabs the information from the user profile"""
-        sql = 'select * from user_table where user_id = %s'
-        cursor = connection.cursor()
-        cursor.execute(sql, [user_id])
-        profile_record = cursor.fetchone()
-        user = User(*profile_record)
-        return user
+        pass
 
     def update_user_profile(self, user: User) -> User:
-        pass
+        """ a method used to update information for the profile besides the image"""
+
+        sql = f"update user_table set user_about = %s, user_birth_date = %s where user_id = %s"
+        cursor = connection.cursor()
+        cursor.execute(sql, user.user_about, user.user_birth_date, user.user_id)
+        connection.commit()
+        return user
 
     def get_user_image(self, user_id: int) -> str:
         """a method to get a user image from the database"""  # need to create a custom exception and database checker
+
+        # Check to see if the post id is in the database, raise an error otherwise.
+        sql = f"select user_id from user_picture_table where user_id = {user_id};"
         cursor = connection.cursor()
+        cursor.execute(sql)
+        if not cursor.fetchone():
+            raise UserImageNotFound('The user image could not be found.')
+
         sql = f"select picture from user_picture_table where user_id = {user_id};"
         cursor.execute(sql)
         image = cursor.fetchone()[0]
@@ -28,8 +37,15 @@ class UserProfileDAOImp(UserProfileDAO):
 
     def update_user_image(self, user_id: int, image: str) -> str:
         """a method to place a user image into the database"""
-        # delete any existing image from the database and place the image in the database
+
+        # Check to see if the user id is in the database, raise an error otherwise.
+        sql = f"select user_id from user_table where user_id = {user_id};"
         cursor = connection.cursor()
+        cursor.execute(sql)
+        if not cursor.fetchone():
+            raise UserNotFound('The user could not be found.')
+
+        # delete any existing image from the database and place the image in the database
         sql = f"DELETE FROM user_picture_table where user_id = {user_id}; " \
               f"INSERT INTO user_picture_table VALUES (default, {user_id}, '{image}');"
         cursor.execute(sql)
@@ -42,10 +58,9 @@ class UserProfileDAOImp(UserProfileDAO):
         image_decoded = image.decode('utf-8')
         return image_decoded
 
-    def update_image_format(self, user_id: int, image_data: str) -> User:
+    def update_user_image_format(self, user_id: int, image_data: str) -> User:
         pass
 
     def update_password(self, user_id: int, password: str) -> User:
         """Stretch"""
         pass
-
